Phase 38 — Type system improvements, string operations, dict operations
Major improvements to type checking and code generation:

Type System:
- Implement type_compatible() for collections with Unknown elements
- Empty dicts/lists now accepted as compatible with typed variants
- Fix BinOp type inference for string concatenation (String + String -> String)
- Add Index expression type inference for dict/list access
- Fix rustc invocation to use ~/.cargo/bin/rustc fallback

String Operations:
- Fix strip/lstrip/rstrip to return String instead of &str
- Implement robust string concatenation using format! macro
- Handle string + string operations with proper type promotion

Dict Operations:
- Fix containment operator for dicts (use contains_key instead of contains)
- Fix dict indexing type inference
- Support dict.pop() with default values

Infrastructure:
- Fix rustc path detection for builds
- Fix test code type annotations (float for division results)

Test Results:
- All 105 type checks passing (fixed test code issues)
- 96/105 builds passing (91.4% success rate)
- 9 remaining issues related to borrow checking and type propagation

+++ examples/data_analysis_tool.py
@@ -31,7 +31,7 @@ def main() -> None:
     print(stats["range"])
 
     # Find above-average months
-    avg: int = stats["sum"] / stats["count"]
+    avg: float = stats["sum"] / stats["count"]
     above_avg: list[int] = [x for x in sales_data if x > avg]
     print(len(above_avg))
 
@@ -55,7 +55,7 @@ def main() -> None:
         growth_rates.append(growth)
 
     total_growth: int = sum(growth_rates)
-    avg_growth: int = total_growth / len(growth_rates)
+    avg_growth: float = total_growth / len(growth_rates)
 
     print(total_growth)
     print(avg_growth)
